chore: rename entity folder to entities
